Move tests out of api.rs
